Conflicts:
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      	app/src/main/java/com/finance/app/view/fragment/loanApplicationFragments/document_checklist/DocumentCheckListFragmentNew.kt
                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      	app/src/main/java/motobeans/architecture/constants/ConstantsApi.kt

+++ app/src/main/java/com/finance/app/persistence/model/HfcPolicyResponse.kt
@@ -15,7 +15,7 @@ class HfcPolicyResponse :Serializable {
     var loanEligibleAmount : Double?=null
     var ltv : Double?=null
     var ltvMutiplier : Double?=null
-    var finalTenure : Float?=null
+    var finalTenure : Int?=null
     var finalEligibility : Double?=null
     var insrValue : Float?=null
     var proposedEMI : Double?=null
